some assets and splash for kelpie (it's meh)

+++ Assets/Scripts/EffectLifetime.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EffectLifetime : MonoBehaviour
 {
-    [SerializeField] private int lifetime = 1;
+    [SerializeField] private float lifetime = 1f;
     void Start()
     {
         Destroy(gameObject, lifetime);
